io-net-tool/backend-django/requirements.txt
# Django and Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Async and WebSockets
channels>=4.0,<5.0
daphne>=4.0,<5.0
channels-redis>=4.1,<5.0

# Background Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# SSH and Remote Execution
paramiko>=3.4,<4.0

# Utilities
python-dotenv>=1.0,<2.0

